combine embedding specs
